requirements.txt
streamlit>=1.30.0
openai>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
watchdog>=3.0.0
supabase>=2.0.0
tenacity>=8.2.0
posthog>=3.0.0
